+++ src/reanimated2/animation/transformationMatrix/TransformMatrix.tsx
@@ -1,4 +1,5 @@
-import type { AffineMatrix, AffineMatrixFlat, Axis } from './matrixUtils';
+import type { Vector3D } from './Vector3D';
+import type { AffineMatrixFlat, Axis } from './matrixUtils';
 import {
   _addMatrices,
   _multiplyMatrices,
@@ -7,7 +8,6 @@ import {
   isAffineMatrix,
   isAffineMatrixFlat,
   _scaleMatrix,
-  unflatten,
   _getRotationMatrix,
 } from './matrixUtils';
 
@@ -19,7 +19,7 @@ function getAngle(angle: number | string) {
   if (typeof angle === 'number') {
     return angle;
   }
-  const angleRegex = /^(?<value>\d+)deg/;
+  const angleRegex = /^(?<value>-?[.\d]+)deg/;
   const numberAsString = angle.match(angleRegex)?.groups?.value;
   if (!numberAsString) {
     throw Error(`Unable to understand angle ${numberAsString}`);
@@ -29,16 +29,16 @@ function getAngle(angle: number | string) {
 }
 
 export class TransformMatrix {
-  value: AffineMatrix;
-  valueFlat: AffineMatrixFlat;
+  public value: number[];
+  private _value: AffineMatrixFlat;
 
   constructor(matrix: Array<Array<number>> | Array<number> | AffineMatrixFlat) {
     if (isAffineMatrix(matrix)) {
-      this.value = matrix;
-      this.valueFlat = flatten(matrix);
+      this._value = flatten(matrix);
+      this.value = flatten(matrix) as number[];
     } else if (isAffineMatrixFlat(matrix)) {
-      this.value = unflatten(matrix);
-      this.valueFlat = matrix;
+      this._value = matrix;
+      this.value = matrix as number[];
     } else {
       throw Error(
         `Cannot concert ${matrix} to TransformMatrix. Please provide an array of length 16 or a nested array 4x4`
@@ -46,41 +46,59 @@ export class TransformMatrix {
     }
   }
 
+  public static getIdentityMatrix() {
+    return new TransformMatrix([
+      ...[1, 0, 0, 0],
+      ...[0, 1, 0, 0],
+      ...[0, 0, 1, 0],
+      ...[0, 0, 0, 1],
+    ]);
+  }
+
   public static getRotationMatrix(angle: number | string, axis: Axis = 'z') {
     const angleValue = getAngle(angle);
-    return _getRotationMatrix(angleValue, axis);
+    return new TransformMatrix(_getRotationMatrix(angleValue, axis));
+  }
+
+  public static getScaleMatrix(scale: number | Vector3D) {
+    const [sx, sy, sz] = Array.isArray(scale) ? scale : [scale, scale, scale];
+
+    return new TransformMatrix([
+      ...[sx, 0, 0, 0],
+      ...[0, sy, 0, 0],
+      ...[0, 0, sz, 0],
+      ...[0, 0, 0, 1],
+    ]);
   }
 
   public static multiplyMatrices = (a: TransformMatrix, b: TransformMatrix) =>
-    new TransformMatrix(_multiplyMatrices(a.valueFlat, b.valueFlat));
+    new TransformMatrix(_multiplyMatrices(a._value, b._value));
 
   public multiply(otherMatrix: TransformMatrix) {
     return new TransformMatrix(
-      _multiplyMatrices(this.valueFlat, otherMatrix.valueFlat)
+      _multiplyMatrices(this._value, otherMatrix._value)
     );
   }
 
   public static addMatrices = (a: TransformMatrix, b: TransformMatrix) => {
-    return new TransformMatrix(_addMatrices(a.valueFlat, b.valueFlat));
+    return new TransformMatrix(_addMatrices(a._value, b._value));
   };
 
   public add(otherMatrix: TransformMatrix) {
-    return new TransformMatrix(
-      _addMatrices(this.valueFlat, otherMatrix.valueFlat)
-    );
+    return new TransformMatrix(_addMatrices(this._value, otherMatrix._value));
   }
 
   public static subtractMatrices = (a: TransformMatrix, b: TransformMatrix) => {
-    return new TransformMatrix(_subtractMatrices(a.valueFlat, b.valueFlat));
+    return new TransformMatrix(_subtractMatrices(a._value, b._value));
   };
 
   public subtract(otherMatrix: TransformMatrix) {
     return new TransformMatrix(
-      _subtractMatrices(this.valueFlat, otherMatrix.valueFlat)
+      _subtractMatrices(this._value, otherMatrix._value)
     );
   }
 
   public scale(scaler: number): TransformMatrix {
-    return new TransformMatrix(_scaleMatrix(this.valueFlat, scaler));
+    return new TransformMatrix(_scaleMatrix(this._value, scaler));
   }
 }
